# Optimize search through output attributes of the target table in merge operation