@Tagar 2021-05-19
@fbad Out of curiosity, are there some performance numbers that show improvements for different number of columns? Thanks