@fbad 2021-05-20
> @fbad Out of curiosity, are there some performance numbers that show improvements for different number of columns? Thanks

Hi @Tagar , no, unfortunately I didn't profile the improvements. Anyway, it might be helpful for you to consider we had a table with 4k columns which basically did not run before this fix (because it was performing that operation, so from Spark UI it looks "hanging" as described by another user in his issue).