@tdas 2021-12-15
Hey fbad, sorry for dropping the ball in this. this is a useful improvement, and I would very like to add it. I think the only thing blocking this is doing some perf benchmark. I can try doing that. Are you still interested in contributing this issue?