@fbad 2021-12-16
@tdas, I wrapped a merge operation (same code in the PR description) in a [scalameter benchmark](https://gist.github.com/fbad/e9febc19e73cc83c36c0dd2b22fe52ee). These are the results:

**cores:** 8
**forked jvm opts**: `-Xmx2g -Xms2g`
**measures for example:** 3 (given the difference in order of magnitude I don't think we need more measures)
**aggregator:** `min`

## Master:

```
Parameters(size -> 1): 2615.426206 ms
Parameters(size -> 10): 2629.348374 ms
Parameters(size -> 100): 4057.518419 ms
Parameters(size -> 300): 38541.363523 ms
Parameters(size -> 1000): stopped
```

## This PR:

```
Parameters(size -> 1): 2792.639716 ms
Parameters(size -> 10): 2309.625942 ms
Parameters(size -> 100): 2496.168179 ms
Parameters(size -> 300): 3522.82363 ms
Parameters(size -> 1000): 12154.483282 ms
```