@scottsand-db 2022-01-06
Hey @fbad, I'm just verifying some performance metrics here. I'm a little unfamiliar with scalameter ... could you please give the command you used to run your benchmark? Thanks!