@weand 2022-07-19
I came across this issue, when looking for a solution for a similar issue.

We are on delta `1.2.1` and try to merge dataframe with a deep (>10 depth) struct of >1000 fields in total. Merge command is just stuck or giving OOM at the end.

As this issue with >1000 columns is available since 1.2.0, I'm wondering, if the still can be another issue with large and deep structs?

Thanks in advance for clarifying.