@Tagar 2022-07-19
@weand are you on Databricks? This was discussed recently on Delta-Users slack - https://delta-users.slack.com/archives/CGK79PLV6/p1656788056366719?thread_ts=1656676461.578359&cid=CGK79PLV6