@weand 2022-07-20
> @weand are you on Databricks? This was discussed recently on Delta-Users slack - https://delta-users.slack.com/archives/CGK79PLV6/p1656788056366719?thread_ts=1656676461.578359&cid=CGK79PLV6 

No, using Delta OSS 1.2.1. Is the improvement also going to be available for OSS?