@Tagar 2022-07-20
@weand that change discussed on Slack is not related to Delta. The fix was in how Parquet files are written, which is a level below Delta