@biancaberdugo 2022-09-30
Just to help someone else with the same issue, seems it was fixed at the version 1.2.0 https://github.com/delta-io/delta/releases/tag/v1.2.0